Fix RIPPLE-83: use proper prepare logic.
Instead of preparing on a GET in the middleware, prepare when the client
reports a new User-Agent is set (either after the bootstrap or when a device is
changed).

Details:

1. Propagate POST to /ripple/user-agent through the middleware.
2. Add cordovaPrepare.js whose sole purpose is to respond to platform changes
   (only for local apps).
3. Remove all the prepare logic from hosted.
