+++ lib/server/emulate.js
@@ -23,6 +23,7 @@ var proxy = require('./proxy'),
     colors = require('colors'),
     express = require('express'),
     cordovaProject = require('./emulate/cordovaProject'),
+    cordovaPrepare = require('./emulate/cordovaPrepare'),
     hosted = require('./emulate/hosted'),
     static = require('./emulate/static');
 
@@ -43,9 +44,7 @@ module.exports = {
         app = proxy.start({route: options.route}, app);
 
         // TODO does not work with custom route (since ripple does not dynamically know custom ones, yet, if set)
-        app.post("/ripple/user-agent", function (req, res/*, next*/) {
-            res.send(200);
-
+        app.post("/ripple/user-agent", function (req, res, next) {
             options.userAgent = unescape(req.body.userAgent);
 
             if (options.userAgent) {
@@ -53,6 +52,7 @@ module.exports = {
             } else {
                 console.log("INFO:".green + ' Using Browser User Agent (String)');
             }
+            next();
         });
 
         // TODO: How to make into a dynamic route (using options.route)? (set at build time right now)
@@ -61,6 +61,7 @@ module.exports = {
         app.use(hosted.inject(options));
 
         if (!options.remote) {
+            app.use("/ripple/user-agent", cordovaPrepare.refresh());
             app.use("/", static.inject(options));
         }
