+++ lib/server/emulate/hosted.js
@@ -21,7 +21,6 @@
 var fs = require('fs'),
     path = require('path'),
     request = require('request'),
-    exec = require('child_process').exec,
     url = require('url'),
 
     HEAD_TAG = /<\s*head\s*>/,
@@ -125,16 +124,6 @@ function remoteInjection(opts) {
 
 function localInjection() {
 
-    var cordovaPrepare;
-
-    try {
-        cordovaPrepare = require.resolve('cordova') ? require('cordova/src/prepare') : null;
-    }
-    catch (e) {
-        console.log('INFO: '.green +
-            'Could not find cordova as a local module. Expecting to find it installed globally.');
-    }
-
     function inject(file, req, res) {
         fs.readFile(file, "utf-8", function (err, data) {
             if (err) { throw new Error(err); }
@@ -172,40 +161,19 @@ function localInjection() {
     }
 
     return function (req, res, next) {
-
-        var callback = function (err) {
-            if (err) {
-                console.error('Failed to execute command: cordova prepare. '.red +
-                    'Make sure you\'ve installed cordova.'.red);
-            }
-
-            // make ripple compatible with phonegap
-            // as of version 3.0 phonegap uses phonegap.js instead of cordova.js
-            // but the files are identical
-            if (req.query.phonegap) {
-                var path = './platforms/' + req.staticPlatform + '/assets/www';
-                fs.readFile(path + '/cordova.js', function(err, data) {
-                    if(err) throw err;
-                    console.log('... copying cordova.js to phonegap.js');
-                    fs.writeFileSync(path + '/phonegap.js', data);
-                });
-            }
-            console.log('... done.');
-            handle(req, res, next);
-        };
-
-        if (req.query.enableripple && req.staticPlatform) {
-            console.log('refreshing project (platform: ' + req.staticPlatform + ') ...');
-            if (!!cordovaPrepare) {
-                cordovaPrepare({ platforms: [ req.staticPlatform ] }).then(callback).done();
-            }
-            else {
-                exec('cordova prepare ' + req.staticPlatform, callback);
-            }
+        // make ripple compatible with phonegap
+        // as of version 3.0 phonegap uses phonegap.js instead of cordova.js
+        // but the files are identical
+        if (req.query.phonegap) {
+            var path = './platforms/' + req.staticPlatform + '/assets/www';
+            fs.readFile(path + '/cordova.js', function(err, data) {
+                if(err) throw err;
+                console.log('... copying cordova.js to phonegap.js');
+                fs.writeFileSync(path + '/phonegap.js', data);
+            });
         }
-        else {
-            handle(req, res, next);
-        }
+
+        handle(req, res, next);
     };
 }
